modifying tiles surface

+++ iOS/G3MiOSSDK/Commons/Basic/CompositeElevationData.hpp
@@ -38,9 +38,8 @@ public:
   void addElevationData(ElevationData* data);
 
   virtual ~CompositeElevationData() {
-    int s = _data.size();
+    size_t s = _data.size();
     for (int i = 0; i < s; i++) {
-#warning CHANGING
       if (_data[i] != NULL){
         _data[i]->_release();
       }
